requirements.txt
# CrewAI
# GROQ
# streamlit

crewai>=0.28.0
langchain>=0.0.335
streamlit>=1.28.0
python-dotenv>=1.0.0
groq>=0.4.0
duckduckgo-search>=3.9.6
google-search-results>=2.4.2
beautifulsoup4>=4.12.2
requests>=2.31.0
pillow>=10.0.0